Separated classes and the interface into separate files to make code more readable and easier to maintain
