+++ MyFirstProgram/IBorrowable.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+
+namespace MyFirstProgram
+{
+    interface IBorrowable
+    {
+        bool IsBorrowed { get; }
+
+        void Borrow(string borrowerName);
+        void Return();
+    }
+}
